Changes15
Admin module updated

+++ src/main/java/com/cg/service/AdminService.java
@@ -3,6 +3,8 @@ package com.cg.service;
 import java.util.List;
 
 import java.util.Optional;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -16,6 +18,8 @@ import org.springframework.stereotype.Service;
 import com.cg.entity.Admin;
 import com.cg.entity.Hall;
 import com.cg.entity.Vendor;
+import com.cg.exception.AdminNotFoundException;
+import com.cg.exception.VendorNotFoundException;
 import com.cg.repository.AdminRepository;
 //import com.cg.repository.HallRepository;
 import com.cg.repository.VendorRepository;
@@ -31,26 +35,99 @@ public class AdminService {
 	@Autowired
 	VendorRepository vendorRepository;
 
-//	//Hall repository instance
-//	@Autowired
-//	HallRepository hallRepository;
-//	
+	Pattern namePattern = Pattern.compile("^[A-Za-z]+$");
 	
 	//Methods for admin
-	public String addAdmin(Admin admin) {
+	public ResponseEntity<Object> addAdmin(Admin admin) {
+		
+		Matcher firstNameMatcher = namePattern.matcher(admin.getFirstName());
+
+		Matcher lastNameMatcher = namePattern.matcher(admin.getLastName());
+		
+		for(char c : admin.getFirstName().toCharArray()){
+			if(Character.isDigit(c))
+	            return new ResponseEntity<Object>("First name should not contain digit.", HttpStatus.BAD_REQUEST);
+		}
+		
+		for(char c : admin.getLastName().toCharArray()){
+			if(Character.isDigit(c))
+	            return new ResponseEntity<Object>("Last name should not contain digit.", HttpStatus.BAD_REQUEST);
+		}
+		
+		if(!firstNameMatcher.matches())
+			return new ResponseEntity<Object>("First name is not valid.", HttpStatus.BAD_REQUEST);
+		
+		if(!lastNameMatcher.matches())
+			return new ResponseEntity<Object>("Last name is not valid.", HttpStatus.BAD_REQUEST);
+		
+		if(admin.getAdminContact().length() != 10)
+			return new ResponseEntity<Object>("Contact number is not valid, must be of 10 digits.", HttpStatus.BAD_REQUEST);
+		
+		for(char c : admin.getAdminContact().toCharArray()){
+			if(!Character.isDigit(c))
+	            return new ResponseEntity<Object>("Contact number is not valid, must be numeric.", HttpStatus.BAD_REQUEST);
+		}
+		
+		String pass = admin.getPassword();
+		if(pass.length() < 8)
+			return new ResponseEntity<Object>("Password should be minimum 8 length.", HttpStatus.BAD_REQUEST);
+		
+		int upperCharCount = 0;
+		int lowerCharCount = 0;
+		int digitCount = 0;
+		int specialCharCount = 0;
+		
+		for(int i=0; i<pass.length(); i++) {
+			
+			if(Character.isUpperCase(pass.charAt(i)))
+				upperCharCount++;
+			
+			if(Character.isLowerCase(pass.charAt(i)))
+				lowerCharCount++;
+			
+			if(Character.isDigit(pass.charAt(i)))
+				digitCount++;
+			
+			if(pass.charAt(i) == '!' || pass.charAt(i) == '@' || 
+				pass.charAt(i) == '#' || pass.charAt(i) == '$' ||
+				pass.charAt(i) == '%' || pass.charAt(i) == '*' )
+				specialCharCount++;
+			
+		}
+		
+		if(upperCharCount == 0)
+			return new ResponseEntity<Object>("Password must contain at least one uppercase character.", HttpStatus.BAD_REQUEST);
+		
+		if(lowerCharCount == 0)
+			return new ResponseEntity<Object>("Password must contain at least one lowercase character.", HttpStatus.BAD_REQUEST);
+		
+		if(digitCount == 0)
+			return new ResponseEntity<Object>("Password must contain at least one digit.", HttpStatus.BAD_REQUEST);
+		
+		if(specialCharCount == 0)
+			return new ResponseEntity<Object>("Password must contain at least one special character from !,@,#,$,%,*.", HttpStatus.BAD_REQUEST);
+		
+		if(upperCharCount + lowerCharCount + digitCount + specialCharCount != pass.length())
+			return new ResponseEntity<Object>("Password does not match the policy, "
+					+ "it should contain at least one uppercase character, lowercase character, digit and special character.", HttpStatus.BAD_REQUEST);
 		
 		adminRepository.save(admin);
 		
-		return "Admin added successfully.";
+		return new ResponseEntity<Object>("Admin added successfully.", HttpStatus.OK);
 	}
 	
 	public String removeAllAdmin() {
 		
-		adminRepository.deleteAll();
-		return "All admin deleted successfully.";
+		if(adminRepository.count() != 0) {
+		
+			adminRepository.deleteAll();
+			return "All admin deleted successfully.";
+		
+		}
+		return "Admin not found.";
 	}
 	
-	public String removeAdminById(int id) {
+	public String removeAdminByAdminId(int id) {
 		
 		if(adminRepository.existsById(id)) {
 			
@@ -59,14 +136,17 @@ public class AdminService {
 		
 		}
 		return "Admin not found.";
+		
 	}
 	
 	public ResponseEntity<Object> getAllAdmin() {
 	
 		List<Admin> admin = adminRepository.findAll();
 		
-		if(admin == null) {
+		if(admin.isEmpty()) {
 			
+			return new ResponseEntity<Object>("Admin not found.", HttpStatus.OK);
+		
 		}
 		return new ResponseEntity<Object>(admin, HttpStatus.OK);
 	}
@@ -79,40 +159,48 @@ public class AdminService {
 
 		if(admin == null) {
 			
+			return new ResponseEntity<Object>("Admin not found.", HttpStatus.OK);
+			
 		}
 		return new ResponseEntity<Object>(admin, HttpStatus.OK);
 	}
 	
-	public ResponseEntity<Object> getAdminById(int id) {
+	public ResponseEntity<Object> getAdminByAdminId(int id) {
 
 		Optional<Admin> admin = adminRepository.findById(id);
 		
 		if(admin == null) {
 			
+			return new ResponseEntity<Object>("Admin not found.", HttpStatus.OK);
+			
 		}
 		return new ResponseEntity<Object>(admin, HttpStatus.OK);
 		
 	}
 	
-	public List<ResponseEntity<Object>> getAdminByFirstName(String firstName){
+	public ResponseEntity<Object> getAdminByFirstName(String firstName){
 		
 		List<Admin> admin = adminRepository.findByFirstName(firstName);
 		
 		if(admin == null ) {
 			
+			return new ResponseEntity<Object>("Admin not found.", HttpStatus.OK);
+			
 		}
-		return (List<ResponseEntity<Object>>) new ResponseEntity<Object>(admin, HttpStatus.OK);
+		return new ResponseEntity<Object>(admin, HttpStatus.OK);
 				
 	}
 	
-	public List<ResponseEntity<Object>> getAdminByLastName(String lastName){
+	public ResponseEntity<Object> getAdminByLastName(String lastName){
 		
 		List<Admin> admin = adminRepository.findByLastName(lastName);
 		
 		if(admin == null ) {
 			
+			return new ResponseEntity<Object>("Admin not found.", HttpStatus.OK);
+			
 		}
-		return (List<ResponseEntity<Object>>) new ResponseEntity<Object>(admin, HttpStatus.OK);
+		return new ResponseEntity<Object>(admin, HttpStatus.OK);
 				
 	}
 	
@@ -121,7 +209,9 @@ public class AdminService {
 		Admin admin = adminRepository.findByAdminContact(adminContact);
 		
 		if(admin == null ) {
-			
+
+			return new ResponseEntity<Object>("Admin not found.", HttpStatus.OK);
+
 		}
 		return new ResponseEntity<Object>(admin, HttpStatus.OK);
 				
@@ -133,29 +223,35 @@ public class AdminService {
 		
 		if(admin == null ) {
 			
+			return new ResponseEntity<Object>("Admin not found.", HttpStatus.OK);
+			
 		}
 		return new ResponseEntity<Object>(admin, HttpStatus.OK);
 				
 	}
 	
-	public List<ResponseEntity<Object>> getAdminSortedByFirstName() {
+	public ResponseEntity<Object> getAdminSortedByFirstName() {
 
 		List<Admin> admin = adminRepository.findAll(Sort.by("firstName"));
 
 		if(admin == null) {
 			
+			return new ResponseEntity<Object>("Admin not found.", HttpStatus.OK);
+			
 		}
-		return (List<ResponseEntity<Object>>) new ResponseEntity<Object>(admin, HttpStatus.OK);
+		return new ResponseEntity<Object>(admin, HttpStatus.OK);
 	}
 	
-	public List<ResponseEntity<Object>> getAdminSortedByLastName() {
+	public ResponseEntity<Object> getAdminSortedByLastName() {
 
 		List<Admin> admin = adminRepository.findAll(Sort.by("lastName"));
 
 		if(admin == null) {
 			
+			return new ResponseEntity<Object>("Admin not found.", HttpStatus.OK);
+			
 		}
-		return (List<ResponseEntity<Object>>) new ResponseEntity<Object>(admin, HttpStatus.OK);
+		return new ResponseEntity<Object>(admin, HttpStatus.OK);
 	}
 	
 	
@@ -173,17 +269,19 @@ public class AdminService {
 		
 		adminRepository.deleteAll();
 		return "All vendor deleted successfully.";
+		
 	}
 	
 	public String removeVendorById(int id) {
 		
-		if(vendorRepository.existsById(id)) {
-			
-			vendorRepository.deleteById(id);
-			return "Vendor deleted successfully.";
+		if(!vendorRepository.existsById(id)) {
+
+			throw new VendorNotFoundException("Vendor not found.");
 		
 		}
-		return "Vendor not found.";
+		
+		vendorRepository.deleteById(id);
+		return "Vendor deleted successfully.";
 	}
 	
 	public ResponseEntity<Object> getAllVendor() {
@@ -191,6 +289,8 @@ public class AdminService {
 		List<Vendor> vendor = vendorRepository.findAll();
 		
 		if(vendor == null) {
+			
+			throw new VendorNotFoundException("Vendor not found.");
 			
 		}
 		return new ResponseEntity<Object>(vendor, HttpStatus.OK);
@@ -204,6 +304,8 @@ public class AdminService {
 
 		if(vendor == null) {
 			
+			throw new VendorNotFoundException("Vendor not found.");
+			
 		}
 		return new ResponseEntity<Object>(vendor, HttpStatus.OK);
 	}
@@ -213,6 +315,8 @@ public class AdminService {
 		Optional<Vendor> vendor= vendorRepository.findById(id);
 		
 		if(vendor == null) {
+			
+			throw new VendorNotFoundException("Vendor not found.");
 			
 		}
 		return new ResponseEntity<Object>(vendor, HttpStatus.OK);
@@ -225,6 +329,8 @@ public class AdminService {
 		
 		if(vendor == null ) {
 			
+			throw new VendorNotFoundException("Vendor not found.");
+			
 		}
 		return (List<ResponseEntity<Object>>) new ResponseEntity<Object>(vendor, HttpStatus.OK);
 				
@@ -236,6 +342,8 @@ public class AdminService {
 		
 		if(vendor == null ) {
 			
+			throw new VendorNotFoundException("Vendor not found.");
+			
 		}
 		return (List<ResponseEntity<Object>>) new ResponseEntity<Object>(vendor, HttpStatus.OK);
 				
@@ -246,6 +354,8 @@ public class AdminService {
 		Vendor vendor = vendorRepository.findByVendorContact(adminContact);
 		
 		if(vendor == null ) {
+			
+			throw new VendorNotFoundException("Vendor not found.");
 			
 		}
 		return new ResponseEntity<Object>(vendor, HttpStatus.OK);
@@ -269,6 +379,8 @@ public class AdminService {
 
 		if(vendor == null) {
 			
+			throw new VendorNotFoundException("Vendor not found.");
+			
 		}
 		return (List<ResponseEntity<Object>>) new ResponseEntity<Object>(vendor, HttpStatus.OK);
 	}
@@ -278,6 +390,8 @@ public class AdminService {
 		List<Vendor> vendor = vendorRepository.findAll(Sort.by("lastName"));
 
 		if(vendor == null) {
+			
+			throw new VendorNotFoundException("Vendor not found.");
 			
 		}
 		return (List<ResponseEntity<Object>>) new ResponseEntity<Object>(vendor, HttpStatus.OK);
